Add a hundred things, one of them being the completed task section
I think I'm back in the game. I'm coming in very clutch Alxamdulillaah. I'm doing a lot more than I thought I could do, and it's not half 3 yet. To think my freedom is this near. But it certainly hasn't come easy. What a struggle it was. The boy who struggle. That would make a good book. And a good boy. The boy who struggled...